delete GUI file, and combined it with MyMainWindow
