+++ vision6D/__init__.py
@@ -30,7 +30,6 @@ LOGGING_CONFIG = {
 logging.config.dictConfig(LOGGING_CONFIG)
 
 from .mainwindow import MyMainWindow
-from .GUI import Interface
 from . import utils
 from . import widgets
 from . import components
