+++ vision6D/entry/main.py
@@ -3,14 +3,14 @@ import os
 import pathlib
 from qtpy import QtWidgets
 
-from .. import Interface
+from .. import MyMainWindow
 
 CWD = pathlib.Path(os.path.abspath(__file__)).parent
 
 def main():
     app = QtWidgets.QApplication(sys.argv)
     with open(CWD.parent / "data" / "style.qss", "r") as f: app.setStyleSheet(f.read())
-    window = Interface()
+    window = MyMainWindow()
     window.showMaximized()
     sys.exit(app.exec_())
